+++ examples/next-app-directory/next.config.mjs
@@ -1,0 +1,11 @@
+/**
+ * @type {import('next').NextConfig}
+ */
+export default {
+  experimental: {
+    appDir: true,
+  },
+  typescript: {
+    tsconfigPath: './tsconfig.nextjs.json',
+  },
+};