+++ examples/next-app-directory/src/builderConfig.js
@@ -1,0 +1,1 @@
+export const API_KEY = 'YOUR API KEY HERE';