+++ examples/vue/vue-2/src/components/HelloWorld.vue
@@ -1,27 +1,44 @@
+<script setup>
+defineProps({
+  msg: {
+    type: String,
+    required: true
+  }
+})
+</script>
+
 <template>
-  <div class="hello-world">
-    <div class="text">Hello {{ text }}!</div>
-    <slot></slot>
+  <div class="greetings">
+    <h1 class="green">{{ msg }}</h1>
+    <h3>
+      You’ve successfully created a project with
+      <a target="_blank" href="https://vitejs.dev/">Vite</a> +
+      <a target="_blank" href="https://v2.vuejs.org/">Vue 2</a>.
+    </h3>
   </div>
 </template>
 
-<script>
-export default {
-  props: {
-    text: {
-      type: String,
-      default: 'world',
-    },
-  },
-};
-</script>
-
 <style scoped>
-.hello-world {
-  display: flex;
-  flex-direction: column;
+h1 {
+  font-weight: 500;
+  font-size: 2.6rem;
+  top: -10px;
 }
-.text {
-  padding: 50px;
+
+h3 {
+  font-size: 1.2rem;
+}
+
+.greetings h1,
+.greetings h3 {
+  text-align: center;
+}
+
+@media (min-width: 1024px) {
+  .greetings h1,
+  .greetings h3 {
+    display: block;
+    text-align: left;
+  }
 }
 </style>