+++ packages/sdks-tests/src/specs/index.ts
@@ -4,7 +4,9 @@ import { CONTENT as homepage } from './homepage.js';
 import { CONTENT as cssNesting } from './css-nesting.js';
 import { CONTENT as symbols, CONTENT_WITHOUT_SYMBOLS } from './symbols.js';
 import { CONTENT as contentBindings } from './content-bindings';
+import { CONTENT as linkUrl } from './link-url.js';
 import { CONTENT as symbolBindings } from './symbol-bindings';
+import { CONTENT as symbolWithInputBinding } from './symbol-with-input-binding.js';
 import { CONTENT as image } from './image.js';
 import { CONTENT as dataBindings } from './data-bindings.js';
 import { CONTENT as dataBindingStyles } from './data-binding-styles.js';
@@ -15,6 +17,7 @@ import {
 import { CONTENT as reactiveState } from './reactive-state';
 import { CONTENT as showHideIf } from './show-hide-if';
 import { CONTENT as textBlock } from './text-block';
+import { CONTENT as stateBinding } from './state-binding';
 import type { BuilderContent } from './types.js';
 
 function isBrowser(): boolean {
@@ -35,6 +38,8 @@ const pages = {
   '/symbols': symbols,
   '/symbols-without-content': CONTENT_WITHOUT_SYMBOLS,
   '/symbol-bindings': symbolBindings,
+  '/link-url': linkUrl,
+  '/symbol-with-input-binding': symbolWithInputBinding,
   '/content-bindings': contentBindings,
   '/image': image,
   '/data-bindings': dataBindings,
@@ -45,6 +50,7 @@ const pages = {
   '/show-hide-if': showHideIf,
   '/custom-breakpoints-reset': customBreakpointsReset,
   '/text-block': textBlock,
+  '/state-binding': stateBinding,
 } as const;
 
 const apiVersionPathToProp = {
@@ -57,8 +63,9 @@ export type Path = keyof typeof pages;
 
 export const ALL_PATHNAMES = Object.keys(pages);
 
-const getContentForPathname = (pathname: string): BuilderContent | null =>
-  pages[pathname as keyof typeof pages] || null;
+const getContentForPathname = (pathname: string): BuilderContent | null => {
+  return pages[pathname as keyof typeof pages] || null;
+};
 
 // remove trailing slash from pathname if it exists
 // unless it's the root path