+++ packages/sdks/e2e/nextjs-react-app-dir/src/app/[[...slug]]/BuilderPage.tsx
@@ -1,0 +1,10 @@
+'use client';
+import { RenderContent } from '@builder.io/sdk-react';
+
+interface BuilderPageProps {
+  builderProps: any;
+}
+
+export default function BuilderPage({ builderProps }: BuilderPageProps) {
+  return <RenderContent {...builderProps} />;
+}