+++ packages/sdks/src/components/render-block/render-repeated-block.lite.tsx
@@ -21,12 +21,14 @@ type Props = {
 export default function RenderRepeatedBlock(props: Props) {
   setContext(BuilderContext, {
     content: props.repeatContext.content,
-    state: props.repeatContext.state,
-    setState: props.repeatContext.setState,
+    localState: props.repeatContext.localState,
+    rootState: props.repeatContext.rootState,
+    rootSetState: props.repeatContext.rootSetState,
     context: props.repeatContext.context,
     apiKey: props.repeatContext.apiKey,
     registeredComponents: props.repeatContext.registeredComponents,
     inheritedStyles: props.repeatContext.inheritedStyles,
+    apiVersion: props.repeatContext.apiVersion,
   });
 
   return <RenderBlock block={props.block} context={props.repeatContext} />;