+++ packages/sdks/src/components/render-content/render-content.lite.tsx
@@ -103,8 +103,8 @@ export default function RenderContent(props: RenderContentProps) {
       data: props.data,
       locale: props.locale,
     }),
-    setContextState: (newState: BuilderRenderState) => {
-      state.contentState = newState;
+    contentSetState: (newRootState: BuilderRenderState) => {
+      state.contentState = newRootState;
     },
 
     allRegisteredComponents: [
@@ -171,7 +171,9 @@ export default function RenderContent(props: RenderContentProps) {
         evaluate({
           code: jsCode,
           context: props.context || {},
-          state: state.contentState,
+          localState: undefined,
+          rootState: state.contentState,
+          rootSetState: state.contentSetState,
         });
       }
     },
@@ -204,7 +206,9 @@ export default function RenderContent(props: RenderContentProps) {
         evaluate({
           code: group,
           context: props.context || {},
-          state: state.contentState,
+          localState: undefined,
+          rootState: state.contentState,
+          rootSetState: state.contentSetState,
         })
       );
     },
@@ -216,7 +220,7 @@ export default function RenderContent(props: RenderContentProps) {
             ...state.contentState,
             [key]: json,
           };
-          state.setContextState(newState);
+          state.contentSetState(newState);
         })
         .catch((err) => {
           console.error('error fetching dynamic data', url, err);
@@ -269,12 +273,14 @@ export default function RenderContent(props: RenderContentProps) {
 
   setContext(builderContext, {
     content: state.useContent,
-    state: state.contentState,
-    setState: state.setContextState,
+    localState: undefined,
+    rootState: state.contentState,
+    rootSetState: TARGET === 'qwik' ? undefined : state.contentSetState,
     context: props.context || {},
     apiKey: props.apiKey,
     apiVersion: props.apiVersion,
     registeredComponents: state.allRegisteredComponents,
+    inheritedStyles: {},
   });
 
   onMount(() => {
@@ -320,7 +326,10 @@ export default function RenderContent(props: RenderContentProps) {
       // override normal content in preview mode
       if (isPreviewing()) {
         const searchParams = new URL(location.href).searchParams;
-        const searchParamPreview = searchParams.get('builder.preview');
+        const searchParamPreviewModel = searchParams.get('builder.preview');
+        const searchParamPreviewId = searchParams.get(
+          `builder.preview.${searchParamPreviewModel}`
+        );
         const previewApiKey =
           searchParams.get('apiKey') || searchParams.get('builder.space');
 
@@ -329,15 +338,14 @@ export default function RenderContent(props: RenderContentProps) {
          * - the preview model name is the same as the one we're rendering, since there can be multiple models rendered
          *  at the same time, e.g. header/page/footer.
          * - the API key is the same, since we don't want to preview content from other organizations.
-         *
-         * TO-DO: should we check that the preview item ID is the same as the initial one being rendered? Or would
-         * this break scenarios where the item is not published yet?
+         * - if there is content, that the preview ID is the same as that of the one we receive.
          *
          * TO-DO: should we only update the state when there is a change?
          **/
         if (
-          searchParamPreview === props.model &&
-          previewApiKey === props.apiKey
+          searchParamPreviewModel === props.model &&
+          previewApiKey === props.apiKey &&
+          (!props.content || searchParamPreviewId === props.content.id)
         ) {
           getContent({
             model: props.model,